starting to refactor profile pages code
